requirements.txt
fastapi
httpx
python-dotenv
python-multipart
uvicorn
typer
pydantic
PyYAML
pytest
Pygments<2.19
pytest-cov
ruff
black
python-docx
pymupdf
openpyxl
streamlit
pre-commit
tqdm
slowapi
prometheus-client
